+++ spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/dto/requestDto/CommentRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.spring_projectclone.dto;
+package com.sparta.spring_projectclone.dto.requestDto;
 
 
 import javax.validation.constraints.NotNull;
