+++ spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/dto/responseDto/CommentResponseDto.java
@@ -1,15 +1,21 @@
 package com.sparta.spring_projectclone.dto.responseDto;
 
+
+import com.sparta.spring_projectclone.model.Comment;
 import lombok.Builder;
-import lombok.Getter;
 
-@Getter
-@Builder
 public class CommentResponseDto {
-
-    private Long commentId;
-    private Long userId;
-    private String userImgUrl;
+    private Long id;
     private String comment;
     private int reviewPoint;
+    private String image;
+
+    @Builder
+    public CommentResponseDto(Comment comment) {
+        this.id = comment.getId();
+        this.comment = comment.getComment();
+        this.reviewPoint = comment.getReviewPoint();
+        this.image = comment.getImage();
+
+    }
 }
