+++ spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/model/Comment.java
@@ -1,7 +1,7 @@
 package com.sparta.spring_projectclone.model;
 
 
-import com.sparta.spring_projectclone.dto.CommentRequestDto;
+import com.sparta.spring_projectclone.dto.requestDto.CommentRequestDto;
 import lombok.Getter;
 import lombok.Setter;
